Fix ServiceHelperExtension::conf
It was using the key argument as both key and value.
